refactor(web): organize devtools per tool
