+++ web/app/components/devtools/react-scan/index.ts
@@ -1,0 +1,2 @@
+export { ReactScanLoader } from './loader'
+export { ReactScan } from './scan'
