+++ web/app/components/devtools/tanstack/index.ts
@@ -1,0 +1,1 @@
+export { TanStackDevtoolsWrapper } from './devtools'
